add some compat layers, move some more things around
