+++ internal/source/testdata/fetch-with-patch.yaml
