+++ internal/source/testdata/git-checkout.yaml
