+++ pkg/config/config.go
@@ -927,15 +927,6 @@ type ReleaseMonitor struct {
 	VersionFilterPrefix string `json:"version-filter-prefix,omitempty" yaml:"version-filter-prefix,omitempty"`
 }
 
-// versionHandler is an interface that defines methods for retrieving version filtering and stripping parameters.
-// It is used to provide a common interface for handling version-related operations for different types of version monitors.
-type versionHandler interface {
-	getStripPrefix() string
-	getStripSuffix() string
-	getFilterPrefix() string
-	getFilterContains() string
-}
-
 // GitHubMonitor indicates using the GitHub API
 type GitHubMonitor struct {
 	// Org/repo for GitHub
@@ -966,66 +957,6 @@ type GitMonitor struct {
 	TagFilterPrefix string `json:"tag-filter-prefix,omitempty" yaml:"tag-filter-prefix,omitempty"`
 	// Filter to apply when searching tags on a GitHub repository
 	TagFilterContains string `json:"tag-filter-contains,omitempty" yaml:"tag-filter-contains,omitempty"`
-}
-
-// GetStripPrefix returns the prefix that should be stripped from the GitMonitor version.
-func (gm *GitMonitor) getStripPrefix() string {
-	return gm.StripPrefix
-}
-
-// GetStripSuffix returns the suffix that should be stripped from the GitMonitor version.
-func (gm *GitMonitor) getStripSuffix() string {
-	return gm.StripSuffix
-}
-
-// GetFilterPrefix returns the prefix filter to apply when searching tags in GitMonitor.
-func (gm *GitMonitor) getFilterPrefix() string {
-	return gm.TagFilterPrefix
-}
-
-// GetFilterContains returns the substring filter to apply when searching tags in GitMonitor.
-func (gm *GitMonitor) getFilterContains() string {
-	return gm.TagFilterContains
-}
-
-// GetStripPrefix returns the prefix that should be stripped from the GitHubMonitor version.
-func (ghm *GitHubMonitor) getStripPrefix() string {
-	return ghm.StripPrefix
-}
-
-// GetStripSuffix returns the suffix that should be stripped from the GitHubMonitor version.
-func (ghm *GitHubMonitor) getStripSuffix() string {
-	return ghm.StripSuffix
-}
-
-// GetFilterPrefix returns the prefix filter to apply when searching tags in GitHubMonitor.
-func (ghm *GitHubMonitor) getFilterPrefix() string {
-	return ghm.TagFilterPrefix
-}
-
-// GetFilterContains returns the substring filter to apply when searching tags in GitHubMonitor.
-func (ghm *GitHubMonitor) getFilterContains() string {
-	return ghm.TagFilterContains
-}
-
-// GetStripPrefix returns the prefix that should be stripped from the ReleaseMonitor version.
-func (rm *ReleaseMonitor) getStripPrefix() string {
-	return rm.StripPrefix
-}
-
-// GetStripSuffix returns the suffix that should be stripped from the ReleaseMonitor version.
-func (rm *ReleaseMonitor) getStripSuffix() string {
-	return rm.StripSuffix
-}
-
-// GetFilterPrefix returns the prefix filter to apply when searching versions in ReleaseMonitor.
-func (rm *ReleaseMonitor) getFilterPrefix() string {
-	return rm.VersionFilterPrefix
-}
-
-// GetFilterContains returns the substring filter to apply when searching versions in ReleaseMonitor.
-func (rm *ReleaseMonitor) getFilterContains() string {
-	return rm.VersionFilterContains
 }
 
 // VersionTransform allows mapping the package version to an APK version
@@ -1138,15 +1069,6 @@ func WithDefaultDisk(disk string) ConfigurationParsingOption {
 func WithDefaultMemory(memory string) ConfigurationParsingOption {
 	return func(options *configOptions) {
 		options.memory = memory
-	}
-}
-
-// withFS sets the fs.FS implementation to use. So far this FS is used only for
-// reading the configuration file. If not provided, the default FS will be an
-// os.DirFS created from the configuration file's containing directory.
-func withFS(filesystem fs.FS) ConfigurationParsingOption {
-	return func(options *configOptions) {
-		options.filesystem = filesystem
 	}
 }
